oracle-devbot/requirements.txt
requests
Pillow
pyautogui
python-telegram-bot==20.7
opencv-python
numpy
langchain
langchain-community
